// types.ts
export interface ImageFile {
  base64: string;
  mimeType: string;
  name: string;
}